requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
google-auth==2.41.0
google-genai==1.39.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.45.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
openai-whisper==20250625
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
torch==2.2.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yt-dlp==2025.9.26
